Move config_template field to RenderConfigMixin
